Add AVR-specific programming connectors
This adds the AVR-ISP-6, AVR-ISP-10, AVR-JTAG-10, AVR-TPI-6 and
AVR-PDI-6 symbols. These replace the symbols by the same name (except
AVR-PDI-6, which was renamed from AVR-PUDI-6) from the old atmel.lib
library.

The actual symbols are completely new, replacing the legacy graphical
representations with a more standard filled box, with pin names on the
inside and pins on the outside. It uses logical grouping of the pins
(VCC at the top, GND at the bottom) instead of following the physical
layout and adds filters for the most commonly used footprints (IDC
header and generic pin header). It also changes the refdes to J? to
comply with the KLC.

The AVR-ISP-6 and AVR-ISP-10 symbols use the same layout, so they are
now drop-in replacements for each other. The AVR-JTAG-10 symbol uses the
same layout as the Conn_ARM_JTAG_SWD_10, so they can be interchanged
(note that the symbol body is slightly different, due to the different
pin length, but pin positions are identical for the pins that are common
to both symbols).