Adding symbols for connectors with mechanical mounting pin.
As discussed in https://github.com/KiCad/kicad-footprints/issues/117

The intention is to have symbols that allows connecting mechanical
mounting pins to a net. This allows connecting the pins to a zone
for increased mechanical strenghts or to allow a zone to better
connect to the area around or under the connector.