Make new KLC script happy by recentering again.
The script originally told me:
Checking symbol 'RJ45_Amphenol_RJMG1BD3B8K1ANR':
  Violating S3.1
    Origin is be centered on the middle of the symbol
    Symbol not centered on origin
     - Center calculated @ (0, -100)

So I moved it, and the script was happy.  Then I updated to the latest
version of the script, and now it says:

  Violating S3.1
    Origin is centered on the middle of the symbol
    Symbol not centered on origin
     - Center calculated @ (0, 100)

So now it wants me to move it back to where I had it originally.