Revert footprint filter hotfix for implicitly added '*'
Previosly we needed to add *: to the footprint filters to force
kicad to use that filter as the start of the footprint name.
(Especially noticable for footprints with single letter prefixes
as their filters resulted in a lot of false positives otherwise.)

A fix has now been commited to the source code that makes this
hotfix obsolete. (It is already in the nightly build of ubuntu)