RFM95W-868S2: Change direction of MISO and MOSI pins.
Previously, MISO had been marked as an input and MOSI as an output.
Since this is a SPI slave device, that is wrong.

The datasheet (page 11) is confused about this subject.  It
(incorrectly) lists MISO as type "I", but then (correctly) describes
it as "SPI Data output".  And vice-versa for MOSI.

http://www.hoperf.com/upload/rf/RFM95_96_97_98W.pdf