CUI PJ-063AH barrel jack
https://www.cui.com/product/resource/digikeypdf/pj-063ah.pdf

This barrel jack does not contain an internal switch, so it only has
two pins, instead of the three that the standard Barrel_Jack symbol
has.

(Actually, since this barrel jack has a metal shield, there ends up
being a third pin for that.  But it does not mean the same thing that
pin 3 means in the standard Barrel_Jack symbol.)

This symbol keeps the same centering as the original Barrel_Jack
symbol, and like the original, it does not have a yellow box around
the symbol.  The KLC script is unhappy about both of these things.