TE 2041021 SD card connector
Datasheet:
http://www.te.com/commerce/DocumentDelivery/DDEController?Action=srchrtrv&DocNm=2041021&DocType=Customer+Drawing&DocLang=English

SD card pinout:
http://pinouts.ru/Memory/sdcard_pinout.shtml

There exists a footprint `Connector_Card:SD_TE_2041021`, but I could
not find a symbol for it.  The existing symbol `SD_Card` has two
"SHELL" pins, but the TE 2041021 only has one "SHELL" pin.

So, this is a copy of the `SD_Card` symbol, with the second "SHELL"
pin removed.

In addition, I changed the electrical type of the DAT pins from Input
to Bidirectional, because Input seemed wrong, according to the pinout
I referenced above.

The KLC script wants the VDD pin at the top, and the VSS pins at the
bottom, but I think that would be weird in this case.  Or at the very
least, it would be a major departure from the existing `SD_Card`
symbol.