Fix footprint associations in Transistor_BJT
Another fix for #520.  This fixes the broken footprint associations in
Transistor_BJT.lib.  MJE13003's datasheet says it has a TO-225 package,
but this is a synonym for TO-126 according to
http://www.onsemi.com/pub/Collateral/AN1040-D.PDF#page=23
so I changed it to TO-126 because that's the name we use in our library.

The pin numbering of MAT02 was badly broken, so I fixed that as well.