Assign one of the existing footprints toCY7C68013A (qfn)
This symbol had no valid footpirnt assigned as the intented footprint
never got merged that way. (wrong EP size)
There was an alternative with top tented vias but it got never assigned
to this symbol.
Now the original footprint has been created using the scripts and
modified to include top tenting as suggested by cypress. (the later is
assigned here)