Add Texas Instruments TPL5110 - Nano Power Timer, Active-Low
This is an ultra low power timer from TI. It draws 35 nA while running.
Timing interval set with a resistor, from 100 ms to 2 h, 1% accuracy
typical. Can be triggered manually. It is designed to be used as a
high-side P-channel MOSFET driver for gating the power to a
microcontroller. It has an optional one-shot mode for enabling the
MOSFET only once.

Datasheet: http://www.ti.com/lit/ds/symlink/tpl5110.pdf

Screenshot:

Footprint: SOT-23-6*
    1. https://github.com/KiCad/kicad-footprints/blob/master/Package_TO_SOT_SMD.pretty/SOT-23-6.kicad_mod
    2. https://github.com/KiCad/kicad-footprints/blob/master/Package_TO_SOT_SMD.pretty/SOT-23-6_Handsoldering.kicad_mod

NOTE: The manufacturer's datasheet differs from the stock KiCad footprint. By the datasheet, the pad centerline distance between the two columns of pads should be 2.7 mm (no tolerance given), but it is only 2.2 mm in the stock KiCad footprint.