TPS2054D: Make IN/OUT pins passive
OUT2 was accidently marked as power input instead of power output, which
resulted in ERC errors with the chips default use case (multi-port USB
current limit). Since the chip could be used for different applications
and is even bidirectional, this patch changes all IN/OUT pins to use the
passive type.

Fixes: #1422