Added dual zeners to Device (#1552)
These symbols are based on the dual schottky diodes and follow the same
naming, but are changed to zener diode graphics. A footprint filter is
also added. Almost all available devices of this kind appear to use some
SOT or SC footprint, so the footprint filter uses a wildcard for those.

These symbols have two warnings from checklib:
 - Symbol not centered on origin
   Due to the graphical nature of the symbol, the center seems a bit
   ambiguous. The centering of this symbol follows the lead of existing
   symbols.
 - Symbol possibly missing default footprint
   This seems a false positive, as this is a generic symbol, so it
   should not have a default footprint.