Driver_LED: Add MPQ2483
This adds the MPQ2483 from Monolithic Power Systems. The datasheet is here:

  https://www.monolithicpower.com/pub/media/document/MPQ2483_r1.05.pdf

One note on the pin arrangement: the FB pin lives on the right side of the
symbol even though it's an input pin, because it is alway fed from the
switching loop that's fed from the BST pin.