Fix checklib.py errors in MCU_Module
This:
 - Sets some pin name offsets to 20mil
 - Removes some symbol names from the description fields
 - Properly centers a symbol

This still leaves missing default footprints for some SODIMM modules.
Currently, there are no SODIMM footprints in the library yet, but even
when there will be, there can very well be different sockets that are
all SODIMM-compatible but with different footprints (effectively, these
modules do not have a footprint of their own, but plug into a socket
that has a footprint).

Note that some SODIMM sockets are being added in
https://github.com/KiCad/kicad-footprints/pull/1187