Fix timing capacitor pin naming of 4538
This fixes the pin names used for the timing capacitors used
by the monostable units (A&B) to be Cx as per the datasheet.
These are connected internally to VSS but they do not need
to be connected externally to anything other than the
timing capacitor.

The symbol outline and center has been changed to match
style guidelines (fixing S3.1 and S3.6 validation failures).

Validation passes apart from S5.1 which is common across
all symbols in this library.