Add Conn_ST_STDC14 (STM32 debug connector, compatible with ARM debug connector)
14-pin connector debug connector. Started from Conn_ARM_JTAG_SWD_10.

Middle 10 pins are the same as ARM Cortex debug connector (except key),
so it is completely compatible with existing cables, connectors, etc.

Extra connections for UART.

`T_` prefix removed from pin names because in my opinion, doesn't add much value,
are harder to read and the symbol would have to be bigger.