Removed inverted input markers from 74173
The datasheet calls the inputs N and M.  They are technically being
inverted inside the logic.  But IMO they should still be called N and M,
without ~, on the symbol. The fact that the interal logic uses the
inverted signals of N and M doesn't mean that the signals are nN and nM.
The only inputs with canonical inverted names are ~G1 and ~G2.

See issue #2367 for details, links and pictures.