Audio:SSI2164: Convert to multi-unit symbol
This commit changes the symbol for the SSI2164 in the Audio library to
use multiple units. The symbol for each VCA unit is modeled after the
symbol used in the datasheet. For the "E" unit (power, mode pin), I had
to improvise a bit; I hope it's adequate.

Running checklib.py on it reports violations of:

S3.1: symbol slightly off-center. This was deliberate, to aim the
Voltage control pin on the VCA units at the center of the triangle
slope. The pin ends are still on the 100mil grid.

S4.1: The script gave no further info, but I can see myself that the pin
lengths are no equal; this, again, is to accomodate the triangular
shape.