Fix active-low input prefixes (#2371)
* Removed inverted input markers from 74173

The datasheet calls the inputs N and M.  They are technically being
inverted inside the logic.  But IMO they should still be called N and M,
without ~, on the symbol. The fact that the interal logic uses the
inverted signals of N and M doesn't mean that the signals are nN and nM.
The only inputs with canonical inverted names are ~G1 and ~G2.

See issue #2367 for details, links and pictures.

* Remove active low prefix on NUCLEO144 (KLC S4.7)

* Remove active low prefix on BNO055 (KLC S4.7)

* Remove active low prefix on ZETA-433-SO (KLC S4.7)

* Remove active low prefix on CY7C6521* (KLC S4.7)

* Remove incorrect active low marker from Ai-Thinker-Ra-01

* Remove active low prefix on MM002 (KLC S4.7)