Add 62256 based on the existing 628128
To create this, I copied the 628128_DIP32_SSOP32 and renamed it 62256_DIP28_SOP28.
I changed the symbol properties to match the new part.
I removed the extra pins, fixed the numbering, and recentered the part.

Like the 628128, there are multiple manufacturers of the 62256:

* [Samsung KM62256C](https://www.futurlec.com/Datasheet/Memory/62256.pdf)
* [Hitachi HM62256B](https://web.mit.edu/6.115/www/document/62256.pdf)
* [Cypress CY62256](https://ecee.colorado.edu/~mcclurel/Cypress_SRAM_CY62256.pdf)

Rather than follow the letter of KLC G2.1 and use a fully specified symbol,
I matched the precedent set by the 628128 and made this a generic symbol.

Similarly, I maintained conformity with the existing 628128 with the I/O pin naming.
The naming varies depending on the manufacturer,
so I used the same names as the 628128.