Fix the control outputs of 8288 for multi-master operation
The 8288 control outputs are incorrectly configured as
outputs. When used in a multi-master configuration this
means it is not possible to connect comand outputs to a bus transceiver such as 74LS245. This is because the resulting
design will fail ERC due to each output being conected to a
tri-state output.

The command outputs should be tri-state. According to the
Command Enable section of the datasheet (page 18-121)  a
logic low on the CEN pin will activate this behaviour.