Fix the control outputs of 8288 for multi-master operation (#2785)
* Fix the control outputs of 8288 for multi-master operation

The 8288 control outputs are incorrectly configured as
outputs. When used in a multi-master configuration this
means it is not possible to connect comand outputs to a bus transceiver such as 74LS245. This is because the resulting
design will fail ERC due to each output being conected to a
tri-state output.

The command outputs should be tri-state. According to the
Command Enable section of the datasheet (page 18-121)  a
logic low on the CEN pin will activate this behaviour.

* Change 8288 INTA command output to tri-state

This changes the INTA pin to be a tri-state output to match
the other Command Outputs on the device.

* Fix footprint filter and pin name offset on 8288

This fixes errors reported by checklib.py. The PDIP footprint
filter appears to be redundant now and no longer matches
any footprint. The pin name offset was being reported as
too long so lower to 20 mils and verify the symbol is still
readable.